ABC.Cを解体sita
