+++ src/calibration/C.C
@@ -1,0 +1,3 @@
+vector<Double_t> C(){
+    return make_C_kev_adc();
+}
